edited

+++ PY4E/temperature.py
@@ -1,3 +1,3 @@
 C = input ("celsius: ")
-F = (int(C) * 1.8)+32
+F = (float(C) * 1.8)+32
 print ("farenhite", F)
